[FIX] move char_to.hpp and update includes
