+++ include/seqan3/range/views/all.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include <seqan3/range/views/char_to.hpp>
+#include <seqan3/alphabet/views/char_to.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/convert.hpp>
 #include <seqan3/range/views/deep.hpp>
