+++ test/unit/range/views/translate_test.cpp
@@ -20,7 +20,7 @@
 #include <seqan3/alphabet/nucleotide/rna5.hpp>
 #include <seqan3/alphabet/nucleotide/rna15.hpp>
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
-#include <seqan3/range/views/char_to.hpp>
+#include <seqan3/alphabet/views/char_to.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/translate.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
