+++ test/unit/search/fm_index_cursor/fm_index_cursor_collection_test.cpp
@@ -13,7 +13,7 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/core/detail/debug_stream_alphabet.hpp>
 #include <seqan3/core/detail/debug_stream_tuple.hpp>
-#include <seqan3/range/views/char_to.hpp>
+#include <seqan3/alphabet/views/char_to.hpp>
 #include <seqan3/search/fm_index/bi_fm_index.hpp>
 #include <seqan3/search/fm_index/bi_fm_index_cursor.hpp>
 #include <seqan3/search/fm_index/fm_index.hpp>
